ebpf: move native custom labels code into native_custom_labels.h
Lift the addr_for_tls_symbol / get_v8_cped_address / native + node
custom-labels readers (lines 161-612 of interpreter_dispatcher.ebpf.c)
into a new header support/ebpf/native_custom_labels.h. The dispatcher
now pulls them in with a single #include placed after the cl_procs map
definition.

Motivation: this ~450-line block of parca-only code sits right next to
maybe_add_apm_info, which is a hot zone for upstream PRs (every change
to the apm-correlation or unwind_stop path conflicts with our block).
Moving the bodies into a brand-new file removes the textual collision
surface entirely while keeping the call sites (maybe_add_native_custom_labels,
maybe_add_node_custom_labels) in place.

No semantic change. Verified by:
- Same instruction count from clang-17 (amd64 44954, arm64 53960);
  the only binary diff is BTF section offsets shifting by 32 bytes
  because the source filename string in debug info is now longer.
- 6.16 distro QEMU test passes: rtld + USDT + cudaverify all green,
  timing events received on first simulation attempt.

The header itself depends on identifiers from bpfdefs.h, tracemgmt.h
(which brings in extmaps.h -> v8_procs), tsd.h, types.h -- all already
included by interpreter_dispatcher.ebpf.c -- plus the cl_procs map
defined directly in the dispatcher. The header is documented with this
contract so future readers know where to look.
